0900-system: use scratch $HOME for `top`
Newer versions of `top` create a $HOME/.config/procps directory if
one does not already exist. Since we are executing as root but with
the calling user's $HOME this can therefore result in a .config
directory being created in the user's home which is owned by root.
This can then cause failures in other tools.

To prevent this we set $HOME to the temporary scratch directory
when executing `top`.

VRVDR-48460