tech-support.functions: Scratch gpg home dir
Prevent any user gpg configuration from affecting the encryption
of tech support output files by setting the gpg home directory to
the temporary scratch directory.

This also prevents a .gnupg being created in the user's home which
is owned by root.

If creation of the scratch directory fails for any reason, and the
user requested encryption of the output, then we print an error and
exit without generating any output. For this reason we also move the
creation of the scratch directory in the tech-support script to
before the "Saving output to ..." message is printed, to clarify that
no output has been generated.

Fixes: VRVDR-49209