Add opt-out shell history collection
In some situations providing technical support might require details
of the commands executed on the device. To achieve this we add the
ability to collect the shell history files of users on the system.

This behaviour is enabled by default when a tech-support report is
generated but can be disabled on a per-report basis using the
"exclude-command-history" option eg.

  $ show tech-support option exclude-command-history
  $ generate tech-support archive option exclude-command-history

As with all tech-support outputs, it is recommended to review the
collected shell histories before passing them to any other party,
removing any information considered sensitive.

Unless the user collecting the report is of the "superuser" level,
only the history of that user is collected for the report. A message
is displayed in the report when this occurs.

VRVDR-51465